+++ src/main/java/com/codeborne/selenide/SelenideConfig.java
@@ -20,6 +20,7 @@ public class SelenideConfig implements Config {
   private boolean driverManagerEnabled = Boolean.parseBoolean(System.getProperty("selenide.driverManagerEnabled", "true"));
   private String browserBinary = System.getProperty("selenide.browserBinary", "");
   private String pageLoadStrategy = System.getProperty("selenide.pageLoadStrategy", "normal");
+  private long pageLoadTimeout = Long.parseLong(System.getProperty("selenide.pageLoadTimeout", "30000"));
   private MutableCapabilities browserCapabilities = new DesiredCapabilities();
 
   private String baseUrl = System.getProperty("selenide.baseUrl", "http://localhost:8080");
@@ -328,8 +329,18 @@ public class SelenideConfig implements Config {
     return pageLoadStrategy;
   }
 
+  @Override
+  public long pageLoadTimeout() {
+    return pageLoadTimeout;
+  }
+
   public SelenideConfig pageLoadStrategy(String pageLoadStrategy) {
     this.pageLoadStrategy = pageLoadStrategy;
+    return this;
+  }
+
+  public SelenideConfig pageLoadTimeout(long pageLoadTimeout) {
+    this.pageLoadTimeout = pageLoadTimeout;
     return this;
   }