+++ statics/src/main/java/com/codeborne/selenide/impl/StaticConfig.java
@@ -162,6 +162,11 @@ public class StaticConfig implements Config {
   }
 
   @Override
+  public long pageLoadTimeout() {
+    return Configuration.pageLoadTimeout;
+  }
+
+  @Override
   public MutableCapabilities browserCapabilities() {
     return Configuration.browserCapabilities;
   }